% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfmcCalc.R
\name{gfmcCalc}
\alias{gfmcCalc}
\title{Grass Fuel Moisture Calculation}
\usage{
gfmcCalc(MC0)
}
\arguments{
\item{MC0}{An output from the mcCalc functions}
}
\value{
\code{gfmc}
}
\description{
This is the actual calculation for grass fuel moisture
}
\seealso{
mcCalc
}
