% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_fuel_moisture_code.r
\name{grass_fuel_moisture_code}
\alias{grass_fuel_moisture_code}
\title{Grass Fuel Moisture Calculation}
\usage{
grass_fuel_moisture_code(MC0)
}
\arguments{
\item{MC0}{An output from the mcCalc functions}
}
\value{
\code{gfmc}
}
\description{
This is the actual calculation for grass fuel moisture
}
\seealso{
mcCalc
}
