% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcCalc.R
\name{mcCalc}
\alias{mcCalc}
\title{Moisture content Calculation}
\usage{
mcCalc(temp, rh, ws, prec, isol, GFMCold, roFL, time.step)
}
\arguments{
\item{temp}{Temperature}

\item{rh}{Relative Humidity}

\item{ws}{Wind Speed}

\item{prec}{Precipitation}

\item{isol}{Insolation}

\item{GFMCold}{Yesterdays Grass Foliar Moisture Content}

\item{roFL}{The nominal fuel load of the fine fuel layer, default is 0.3
kg/m^2}

\item{time.step}{Time step (hour) [default 1 hour]}
}
\description{
Calculation of moisture content for use int he GFMC calculation
}
\seealso{
\code{gfmc}
}
