\name{test_fbp}
\docType{data}
\alias{test_fbp}
\title{Fire Behaviour Prediction Sample Data Set}
\description{
  This data set is a set of input data for each of the test cases in the publication supplied below.
}
\usage{test_fbp}
\format{A data frame containing 24 columns, 21 rows, including 1 header line}
\source{
   \url{http://cfs.nrcan.gc.ca/pubwarehouse/pdfs/31414.pdf}
}
\references{
1. Wotton, B.M., Alexander, M.E., Taylor, S.W. 2009. Updates and revisions to the 1992 Canadian forest fire behavior prediction system. Nat. Resour. Can., Can. For. Serv., Great Lakes For. Cent., Sault Ste. Marie, Ontario, Canada. Information Report GLC-X-10, 45p.
}
\keyword{datasets}
