% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cffdrs-package.R
\docType{data}
\name{test_fbpRaster}
\alias{test_fbpRaster}
\title{Raster Data for fbpRaster function}
\format{
A raster (tif) file.
}
\description{
Test raster file to calculate fbp data.
}
\keyword{datasets}
