\name{test_gfmc}
\docType{data}
\alias{test_gfmc}
\title{Grass Fuel Moisture Code Sample Input Data Set}
\description{
  This data set is the sample input data that was used in original FWI program calibration.
}
\usage{data(test_gfmc)}
\format{A data frame containing 9 columns and 199 rows, with 1 header line}
\keyword{datasets}
