\name{test_hffmc}
\docType{data}
\alias{test_hffmc}
\title{Hourly Fine Fuel Moisture Code Sample Input Data Set}
\description{
  Sample dataset for use with the \code{hffmc} function.
}
\usage{data(test_hffmc)}
\format{A data frame containing 8 columns and 481 rows, including 1 header line}
\keyword{datasets}
