\name{test_lros}
\docType{data}
\alias{test_lros}
\title{Line-based Simard function Sample Data Set}
\description{
  This is a set of input data to test the lros function.
}
\usage{test_lros}
\format{A data frame containing 8 columns, 4 rows, including 1 header line. }
\source{
  no source
}
\references{
  none
}
\keyword{datasets, simard, lros}
