\name{test_pros}
\docType{data}
\alias{test_pros}
\title{Point-based Simard function Sample Data Set}
\description{
  This is a set of input data to test the pros function.
}
\usage{test_pros}
\format{A data frame containing 9 columns, 4 rows, including 1 header line.}
\source{
  no source
}
\references{
  none
}
\keyword{datasets, simard, lros}
