\name{test_rast_hour02}
\docType{data}
\alias{test_rast_hour02}
\title{Raster Data for ffmcRaster function}
\description{
 Hourly fire weather inputs obtained from the Global Environmental Multiscale Model (GEM) in northern Alberta
}
\usage{test_rast_hour02}
\format{A raster (tif) file.}
\keyword{datasets}
