\name{test_sdmc}
\docType{data}
\alias{test_sdmc}
\title{Sheltered Duff Moisture Code Sample Input Data Set}
\description{
  This data set is the sample input data that was used in original FWI program calibration, but with an initial dmc value populated.
}
\usage{test_fwi}
\format{A data frame containing 10 columns and 49 rows, including 1 header line}
\source{
\url{http://cfs.nrcan.gc.ca/pubwarehouse/pdfs/19973.pdf}
}
\references{
1.	Van Wagner, CE. and T.L. Pickett. 1985. Equations and FORTRAN program for the Canadian Forest Fire Weather Index System. Can. For. Serv., Ottawa, Ont. For. Tech. Rep. 33. 18 p.
}
\keyword{datasets}
