\name{test_wDC}
\docType{data}
\alias{test_wDC}
\title{Overwinter Drought Code Sample Input Data Set}
\description{
  This dataset has 2 ID values (weather stations), and each have 2 sequential years. This data can be used as an example to calculated overwintered DC. There are 10 columns and 1463 rows, including 1 header row.
}
\usage{data(test_wDC)}
\format{A data frame containing 10 columns and 1463 rows, including 1 header line}
\keyword{datasets}
