\name{test_wDC_fs}
\docType{data}
\alias{test_wDC_fs}
\title{Fire Season Dataset to test Overwinter Drought Code}
\description{
  This dataset has pre-set start and end dates to the fire season for 2 weather stations. The point of this dataset is to demonstrate that a data frame of start and
  end dates for the fire season can be calculated and applied to the program.
}
\usage{test_wDC_fs}
\format{A data frame containing 7 columns and 9 rows, including 1 header line}
\keyword{datasets}
