% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cffdrs-package.R
\docType{data}
\name{test_fwi}
\alias{test_fwi}
\title{Fire Weather Index Sample Input Data Set}
\format{
A data frame containing 9 columns and 49 rows, with 1 header line
}
\source{
\url{http://cfs.nrcan.gc.ca/pubwarehouse/pdfs/19973.pdf}
}
\description{
This data set is the sample input data that was used in original FWI program
calibration.
}
\references{
1. Van Wagner, CE. and T.L. Pickett. 1985. Equations and FORTRAN
program for the Canadian Forest Fire Weather Index System. Can. For. Serv.,
Ottawa, Ont. For. Tech. Rep. 33. 18 p.
}
\keyword{datasets}
