% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cffdrs-package.R
\docType{data}
\name{test_gfmc}
\alias{test_gfmc}
\title{Grass Fuel Moisture Code Sample Input Data Set}
\format{
A data frame containing 9 columns and 199 rows, with 1 header line
}
\description{
This data set is the sample input data that was used in original FWI program
calibration.
}
\keyword{datasets}
