% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cffdrs-package.R
\docType{data}
\name{test_hffmc}
\alias{test_hffmc}
\title{Hourly Fine Fuel Moisture Code Sample Input Data Set}
\format{
A data frame containing 8 columns and 481 rows, including 1 header
line
}
\description{
Sample dataset for use with the \code{hffmc} function.
}
\keyword{datasets}
