% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cffdrs-package.R
\docType{data}
\name{test_rast_day02}
\alias{test_rast_day02}
\title{Raster Data for fwiRaster function}
\format{
A raster (tif) file.
}
\description{
Daily fire weather inputs obtained from the Global Environmental Multiscale
Model (GEM) in northern Alberta
}
\keyword{datasets}
