% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cffdrs-package.R
\docType{data}
\name{test_wDC}
\alias{test_wDC}
\title{Overwinter Drought Code Sample Input Data Set}
\format{
A data frame containing 10 columns and 1463 rows, including 1 header
line
}
\description{
This dataset has 2 ID values (weather stations), and each have 2 sequential
years. This data can be used as an example to calculated overwintered DC.
There are 10 columns and 1463 rows, including 1 header row.
}
\keyword{datasets}
